+++ AWSCore/Service/AWSService.m
@@ -293,6 +293,7 @@ static NSString *const AWSServiceNameComprehend = @"comprehend";
 static NSString *const AWSServiceNameKinesisVideo = @"kinesisvideo";
 static NSString *const AWSServiceNameKinesisVideoArchivedMedia = @"kinesisvideo";
 static NSString *const AWSServiceNameSageMakerRuntime = @"sagemaker";
+static NSString *const AWSServiceNameTranscribeStreaming = @"transcribe";
 
 @interface AWSEndpoint()
 
@@ -489,6 +490,8 @@ static NSString *const AWSServiceNameSageMakerRuntime = @"sagemaker";
             return AWSServiceNameCognitoIdentityProvider;
         case AWSServiceCognitoSync:
             return AWSServiceNameCognitoSync;
+        case AWSServiceComprehend:
+            return AWSServiceNameComprehend;
         case AWSServiceConnect:
             return AWSServiceNameConnect;
         case AWSServiceDynamoDB:
@@ -497,16 +500,20 @@ static NSString *const AWSServiceNameSageMakerRuntime = @"sagemaker";
             return AWSServiceNameEC2;
         case AWSServiceElasticLoadBalancing:
             return AWSServiceNameElasticLoadBalancing;
+        case AWSServiceFirehose:
+            return AWSServiceNameFirehose;
         case AWSServiceIoT:
             return AWSServiceNameIoT;
         case AWSServiceIoTData:
             return AWSServiceNameIoTData;
-        case AWSServiceFirehose:
-            return AWSServiceNameFirehose;
-        case AWSServiceKinesis:
-            return AWSServiceNameKinesis;
         case AWSServiceKMS:
             return AWSServiceNameKMS;
+        case AWSServiceKinesis:
+            return AWSServiceNameKinesis;
+        case AWSServiceKinesisVideo:
+            return AWSServiceNameKinesisVideo;
+        case AWSServiceKinesisVideoArchivedMedia:
+            return AWSServiceNameKinesisVideoArchivedMedia;
         case AWSServiceLambda:
             return AWSServiceNameLambda;
         case AWSServiceLexRuntime:
@@ -517,38 +524,34 @@ static NSString *const AWSServiceNameSageMakerRuntime = @"sagemaker";
             return AWSServiceNameMachineLearning;
         case AWSServiceMobileAnalytics:
             return AWSServiceNameMobileAnalytics;
-        case AWSServicePolly:
-            return AWSServiceNamePolly;
         case AWSServiceMobileTargeting:
             return AWSServiceNameMobileTargeting;
+        case AWSServicePolly:
+            return AWSServiceNamePolly;
         case AWSServiceRekognition:
             return AWSServiceNameRekognition;
         case AWSServiceS3:
             return AWSServiceNameS3;
         case AWSServiceSES:
             return AWSServiceNameSES;
-        case AWSServiceSimpleDB:
-            return AWSServiceNameSimpleDB;
         case AWSServiceSNS:
             return AWSServiceNameSNS;
         case AWSServiceSQS:
             return AWSServiceNameSQS;
         case AWSServiceSTS:
             return AWSServiceNameSTS;
+        case AWSServiceSageMakerRuntime:
+            return AWSServiceNameSageMakerRuntime;
+        case AWSServiceSimpleDB:
+            return AWSServiceNameSimpleDB;
         case AWSServiceTextract:
             return AWSServiceNameTextract;
         case AWSServiceTranscribe:
             return AWSServiceNameTranscribe;
+        case AWSServiceTranscribeStreaming:
+            return AWSServiceNameTranscribeStreaming;
         case AWSServiceTranslate:
             return AWSServiceNameTranslate;
-        case AWSServiceComprehend:
-            return AWSServiceNameComprehend;
-        case AWSServiceKinesisVideo:
-            return AWSServiceNameKinesisVideo;
-        case AWSServiceKinesisVideoArchivedMedia:
-            return AWSServiceNameKinesisVideoArchivedMedia;
-        case AWSServiceSageMakerRuntime:
-            return AWSServiceNameSageMakerRuntime;
         default:
             return nil;
     }
@@ -635,6 +638,8 @@ static NSString *const AWSServiceNameSageMakerRuntime = @"sagemaker";
         URL = [NSURL URLWithString:[NSString stringWithFormat:@"%@://pinpoint%@%@.amazonaws.com", HTTPType, separator, regionName]];
     } else if (serviceType == AWSServiceSageMakerRuntime) {
         URL = [NSURL URLWithString:[NSString stringWithFormat:@"%@://runtime.%@%@%@.amazonaws.com", HTTPType, serviceName, separator, regionName]];
+    } else if (serviceType == AWSServiceTranscribeStreaming) {
+        URL = [NSURL URLWithString:[NSString stringWithFormat:@"%@://transcribestreaming%@%@.amazonaws.com", HTTPType, separator, regionName]];
     } else {
         URL = [NSURL URLWithString:[NSString stringWithFormat:@"%@://%@%@%@.amazonaws.com", HTTPType, serviceName, separator, regionName]];
     }